Refactor JS: 1 file -> 1 component
