+++ src/js/components/footer.js
@@ -4,8 +4,11 @@ import * as React from 'react';
 
 const Footer = (props: { logoSrc: string }) => (
   <footer
-    className="slds-grid slds-grid--align-spread
-      site-contentinfo slds-grid_vertical-align-center spacing-x-large"
+    className="slds-grid
+      slds-grid--align-spread
+      site-contentinfo
+      slds-grid_vertical-align-center
+      spacing-x-large"
   >
     <div
       className="footer-logo"
